clean code
- deleted celery & redis related code
- update docker-compose & docker-file (had to move `pip install` to run.sh because it is not work on `Dockerfile`)
