+++ backend/.gitignore
@@ -1,3 +1,6 @@
 __pycache__
 app.egg-info
 .DS_Store
+
+# pyenv
+./env-scdao
