+++ backend/Dockerfile
@@ -1,16 +1,11 @@
-FROM python:3.8
+FROM python:3.10
+
+WORKDIR /app
 
 ENV PYTHONUNBUFFERED 1
 
-RUN mkdir /app
-WORKDIR /app
-
 RUN apt update && \
-    apt install -y postgresql-client && \
-    apt install -y tesseract-ocr
+    apt install -y postgresql-client
 
-
-COPY requirements.txt ./
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
+COPY requirements.txt  /app
+RUN pip install -r requirements.txt
