+++ backend/app/core/security.py
@@ -1,8 +1,5 @@
-import time
-
 import jwt
 from fastapi.security import OAuth2PasswordBearer
-from fastapi import Request, HTTPException
 from passlib.context import CryptContext
 from datetime import datetime, timedelta
