+++ backend/app/crud/cc_crud.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException
 from sqlalchemy.orm import Session
 from starlette import status
 
-from app.schemas import cc_schemas, user_schemas
+from app.schemas import cc_schemas 
 import app.models.form_models as form_models
