+++ backend/conftest.py
@@ -9,14 +9,12 @@ from fastapi.testclient import TestClient
 from fastapi import Request
 import typing as t
 
-from app.celery_system.task_classes import TaskStates, TaskTypes
 from app.core import config, security
 from app.core.security import user_over_rate_limit
 from app.db.session import get_db
 from app.db.base_class import Base
 from app import models
 from app.main import app
-from celery.backends.database import models as celery_models
 
 
 def get_test_db_url() -> str:
@@ -243,31 +241,3 @@ def county_authorized_token_headers(
     a_token = tokens["access_token"]
     headers = {"Authorization": f"Bearer {a_token}"}
     return headers
-
-
-###############################################
-## Task API Tests ############################
-
-
-@pytest.fixture
-def test_celery_task_record(test_db) -> celery_models.Task:
-    task = celery_models.Task(
-        task_id=str(uuid4()))
-    task.status = TaskStates.SUCCESS
-    task.result = pickle.dumps({'id': 1, 'type': TaskTypes.CCF})
-    test_db.add(task)
-    test_db.commit()
-    test_db.refresh(task)
-    return task
-
-
-@pytest.fixture
-def test_celery_task_record_no_result(test_db) -> celery_models.Task:
-    task = celery_models.Task(
-        task_id=str(uuid4()))
-    task.status = TaskStates.SUCCESS
-    task.result = bytes()
-    test_db.add(task)
-    test_db.commit()
-    test_db.refresh(task)
-    return task
