+++ backend/requirements.txt
@@ -1,21 +1,51 @@
-alembic==1.4.2
-Authlib==0.14.3
-fastapi==0.78.0
-httpx==0.23.0
-ipython==8.4.0
-itsdangerous==1.1.0
-Jinja2==2.11.2
-psycopg2==2.9.3
-pytest==6.0.1
-requests==2.24.0
-SQLAlchemy==1.3.19
-starlette==0.19.1
-uvicorn==0.17.6
-passlib==1.7.2
-bcrypt==3.1.7
-sqlalchemy-utils==0.36.8
-python-multipart==0.0.5
-pyjwt==1.7.1
-Pillow==9.1.1
-boto3==1.14.43
-google-cloud-documentai==1.5.0
+alembic==1.8.1
+anyio==3.6.1
+attrs==22.1.0
+boto3==1.24.90
+botocore==1.27.90
+cachetools==5.2.0
+certifi==2022.9.24
+cffi==1.15.1
+charset-normalizer==2.1.1
+click==8.1.3
+cryptography==38.0.1
+fastapi==0.85.0
+google-api-core==2.10.2
+google-auth==2.12.0
+google-cloud-documentai==2.0.3
+googleapis-common-protos==1.56.4
+grpcio==1.49.1
+grpcio-status==1.49.1
+h11==0.14.0
+idna==3.4
+iniconfig==1.1.1
+jmespath==1.0.1
+Mako==1.2.3
+MarkupSafe==2.1.1
+packaging==21.3
+passlib==1.7.4
+pluggy==1.0.0
+proto-plus==1.22.1
+protobuf==4.21.7
+py==1.11.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.21
+pydantic==1.10.2
+PyJWT==2.5.0
+pyparsing==3.0.9
+pytest==7.1.3
+python-dateutil==2.8.2
+python-dotenv==0.21.0
+requests==2.28.1
+rsa==4.9
+s3transfer==0.6.0
+six==1.16.0
+sniffio==1.3.0
+SQLAlchemy==1.4.41
+SQLAlchemy-Utils==0.38.3
+starlette==0.20.4
+tomli==2.0.1
+typing_extensions==4.4.0
+urllib3==1.26.12
+uvicorn==0.18.3
