+++ backend/run.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+pip install -r requirements.txt
+
+python3 app/main.py
