+++ docker-compose.override.yml
@@ -15,10 +15,8 @@ services:
   backend:
     volumes:
       - ./backend:/app/:cached
-      - ./.docker/.ipython:/root/.ipython:cached
     depends_on:
       - "postgres"
-
   frontend:
     environment:
       NODE_ENV: 'production'
